% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications-codebooks.R
\name{alprek_applications_funding_types}
\alias{alprek_applications_funding_types}
\title{Load Applications Funding Types}
\usage{
alprek_applications_funding_types()
}
\value{
A tibble with columns: \code{funding_type}, \code{funding_category}, \code{notes}.
}
\description{
Returns the mapping from ADECE funding type labels to
standardized funding categories.
}
\examples{
alprek_applications_funding_types()

}
