% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helpers.R
\name{alprek_clean_colnames}
\alias{alprek_clean_colnames}
\title{Clean column names by squishing whitespace}
\usage{
alprek_clean_colnames(x)
}
\arguments{
\item{x}{Character vector of column names.}
}
\value{
Character vector with squeezed whitespace.
}
\description{
Clean column names by squishing whitespace
}
\keyword{internal}
