% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget-codebooks.R
\name{alprek_delivery_types}
\alias{alprek_delivery_types}
\title{Load Delivery Type Codes}
\usage{
alprek_delivery_types()
}
\value{
A tibble with columns: \code{code}, \code{name}, \code{name_short}.
}
\description{
Returns the mapping from single-letter delivery type codes
to full names.
}
\examples{
alprek_delivery_types()

}
