% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-codebooks.R
\name{alprek_race_mapping}
\alias{alprek_race_mapping}
\title{Load Race/Ethnicity Mapping}
\usage{
alprek_race_mapping()
}
\value{
A tibble with columns: \code{raw_value}, \code{standardized}, \code{factor_order}.
}
\description{
Returns the mapping from raw race/ethnicity text values to
standardized categories. Maps diverse raw entries (e.g., "Black or African
American", "Black/African American") to a consistent set of 7 categories.
}
\details{
Standardized categories and their factor order:
\enumerate{
\item White
\item Black
\item Latino/Hispanic
\item Asian
\item Mixed
\item Other
\item Unknown
}
}
\examples{
alprek_race_mapping()

}
