% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helpers.R
\name{alprek_school_year_to_start}
\alias{alprek_school_year_to_start}
\title{Convert School Year to Start Year}
\usage{
alprek_school_year_to_start(school_year)
}
\arguments{
\item{school_year}{Character. School year in "YYYY-YYYY" format.}
}
\value{
Integer start year.
}
\description{
Extracts the first year from a "YYYY-YYYY" school year string.
}
\examples{
alprek_school_year_to_start("2024-2025")
# 2024L

}
