% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget-master.R
\name{budget_summary_stats}
\alias{budget_summary_stats}
\title{Calculate Summary Statistics by Group}
\usage{
budget_summary_stats(x, by = "school_year")
}
\arguments{
\item{x}{An \code{alprek_budget_master} or \code{alprek_budget_panel} object.}

\item{by}{Character vector of grouping variable names.
Default is \code{"school_year"}.}
}
\value{
A tibble with summary statistics.
}
\description{
Calculates descriptive statistics for budget data grouped by
one or more variables.
}
