% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-export.R
\name{classroom_export_csv}
\alias{classroom_export_csv}
\title{Export Classroom Data to CSV}
\usage{
classroom_export_csv(x, path)
}
\arguments{
\item{x}{An \code{alprek_classroom_clean} or \code{alprek_classroom_panel} object.}

\item{path}{Character. Output file path.}
}
\value{
Invisible file path of the written file.
}
\description{
Exports processed classroom data to CSV format.
}
