% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-export.R
\name{classroom_export_parquet}
\alias{classroom_export_parquet}
\title{Export Classroom Data to Parquet}
\usage{
classroom_export_parquet(x, path, compression = "snappy")
}
\arguments{
\item{x}{An \code{alprek_classroom_clean} or \code{alprek_classroom_panel} object.}

\item{path}{Character. Output file path.}

\item{compression}{Character. Compression algorithm. Default \code{"snappy"}.}
}
\value{
Invisible file path.
}
\description{
Exports processed classroom data to Parquet format.
Requires the \code{arrow} package.
}
