% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-export.R
\name{classroom_export_rds}
\alias{classroom_export_rds}
\title{Export Classroom Data to RDS}
\usage{
classroom_export_rds(x, path, compress = TRUE)
}
\arguments{
\item{x}{An \code{alprek_classroom_clean} or \code{alprek_classroom_panel} object.}

\item{path}{Character. Output file path.}

\item{compress}{Logical. Use compression? Default \code{TRUE}.}
}
\value{
Invisible file path.
}
\description{
Exports processed classroom data to R's native serialized format.
}
