% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-core.R
\name{db_close}
\alias{db_close}
\title{Close ALprekDB DuckDB Connection}
\usage{
db_close(conn)
}
\arguments{
\item{conn}{A DBI connection object from \code{\link[=db_init]{db_init()}}.}
}
\value{
Invisible \code{TRUE} on success.
}
\description{
Properly closes the DuckDB database connection and shuts down
the database driver.
}
