% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget-read.R
\name{.budget_check_input_path}
\alias{.budget_check_input_path}
\title{Check budget input path for known noncanonical source files}
\usage{
.budget_check_input_path(path)
}
\description{
Check budget input path for known noncanonical source files
}
\keyword{internal}
