% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-clean.R
\name{.clean_administrative_vars}
\alias{.clean_administrative_vars}
\title{Clean administrative variables}
\usage{
.clean_administrative_vars(df)
}
\description{
Clean administrative variables
}
\keyword{internal}
