% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-clean.R
\name{.clean_demographic_flags}
\alias{.clean_demographic_flags}
\title{Clean demographic flag variables}
\usage{
.clean_demographic_flags(df)
}
\description{
Clean demographic flag variables
}
\keyword{internal}
