% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-clean.R
\name{.clean_fluent_language}
\alias{.clean_fluent_language}
\title{Clean fluent language field}
\usage{
.clean_fluent_language(raw_lang, language_mapping)
}
\description{
Clean fluent language field
}
\keyword{internal}
