% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-clean.R
\name{.clean_iep_referral}
\alias{.clean_iep_referral}
\title{Clean IEP referral variables and derive iep2}
\usage{
.clean_iep_referral(df)
}
\description{
Clean IEP referral variables and derive iep2
}
\keyword{internal}
