% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-clean.R
\name{.clean_staff_contacts}
\alias{.clean_staff_contacts}
\title{Combine first name + last name into single name columns for staff}
\usage{
.clean_staff_contacts(df)
}
\description{
Combine first name + last name into single name columns for staff
}
\keyword{internal}
