% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-clean.R
\name{.clean_student_delivery_type}
\alias{.clean_student_delivery_type}
\title{Clean delivery type using CSV mapping}
\usage{
.clean_student_delivery_type(df, delivery_mapping)
}
\description{
Clean delivery type using CSV mapping
}
\keyword{internal}
