% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-clean.R
\name{.clean_student_demographics}
\alias{.clean_student_demographics}
\title{Clean student demographics}
\usage{
.clean_student_demographics(df, race_mapping, school_year)
}
\description{
Clean student demographics
}
\keyword{internal}
