% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-clean.R
\name{.clean_student_staff_contacts}
\alias{.clean_student_staff_contacts}
\title{Clean staff contacts (new format only)}
\usage{
.clean_student_staff_contacts(df)
}
\description{
Clean staff contacts (new format only)
}
\keyword{internal}
