% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-clean.R
\name{.concat_name}
\alias{.concat_name}
\title{Concatenate first and last name, handling NA}
\usage{
.concat_name(fname, lname)
}
\description{
Concatenate first and last name, handling NA
}
\keyword{internal}
