% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-clean.R
\name{.concat_student_name}
\alias{.concat_student_name}
\title{Concatenate first and last name, handling NAs}
\usage{
.concat_student_name(first, last)
}
\description{
Concatenate first and last name, handling NAs
}
\keyword{internal}
