% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-read.R
\name{.db_build_classroom_budget_diagnostics}
\alias{.db_build_classroom_budget_diagnostics}
\title{Reconstruct classroom-budget diagnostics from master classroom table}
\usage{
.db_build_classroom_budget_diagnostics(classroom_df)
}
\description{
Reconstruct classroom-budget diagnostics from master classroom table
}
\keyword{internal}
