% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-read.R
\name{.db_build_master_diagnostics}
\alias{.db_build_master_diagnostics}
\title{Reconstruct master diagnostics from DuckDB master tables}
\usage{
.db_build_master_diagnostics(classroom_df, student_df)
}
\description{
Reconstruct master diagnostics from DuckDB master tables
}
\keyword{internal}
