% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-read.R
\name{.db_build_panel_object}
\alias{.db_build_panel_object}
\title{Build panel S3 object from data frame}
\usage{
.db_build_panel_object(df, module)
}
\description{
Build panel S3 object from data frame
}
\keyword{internal}
