% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-read.R
\name{.db_build_student_classroom_diagnostics}
\alias{.db_build_student_classroom_diagnostics}
\title{Reconstruct student-classroom diagnostics from DuckDB master tables}
\usage{
.db_build_student_classroom_diagnostics(classroom_df, student_df)
}
\description{
Reconstruct student-classroom diagnostics from DuckDB master tables
}
\keyword{internal}
