% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-schema.R
\name{.db_create_schema}
\alias{.db_create_schema}
\title{Create database schema (metadata + column type registry tables)}
\usage{
.db_create_schema(conn)
}
\description{
Create database schema (metadata + column type registry tables)
}
\keyword{internal}
