% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-write.R
\name{.db_existing_years}
\alias{.db_existing_years}
\title{Get existing school_year values from a table}
\usage{
.db_existing_years(conn, table)
}
\description{
Get existing school_year values from a table
}
\keyword{internal}
