% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-write.R
\name{.db_extract_panel_data}
\alias{.db_extract_panel_data}
\title{Extract data frame from panel object}
\usage{
.db_extract_panel_data(obj)
}
\description{
Extract data frame from panel object
}
\keyword{internal}
