% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-schema.R
\name{.db_get_column_types}
\alias{.db_get_column_types}
\title{Get column type registry for a table}
\usage{
.db_get_column_types(conn, table_name)
}
\description{
Get column type registry for a table
}
\keyword{internal}
