% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-read.R
\name{.db_master_budget_years}
\alias{.db_master_budget_years}
\title{Infer budget-covered years from a master classroom table}
\usage{
.db_master_budget_years(classroom_df)
}
\description{
Infer budget-covered years from a master classroom table
}
\keyword{internal}
