% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-read.R
\name{.db_master_has_budget}
\alias{.db_master_has_budget}
\title{Determine rows with reconstructed budget data}
\usage{
.db_master_has_budget(classroom_df)
}
\description{
Determine rows with reconstructed budget data
}
\keyword{internal}
