% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-schema.R
\name{.db_merge_factor_levels}
\alias{.db_merge_factor_levels}
\title{Merge factor levels (union of old and new)}
\usage{
.db_merge_factor_levels(conn, table_name, df)
}
\description{
Merge factor levels (union of old and new)
}
\keyword{internal}
