% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-schema.R
\name{.db_validate_schema}
\alias{.db_validate_schema}
\title{Validate existing database schema version}
\usage{
.db_validate_schema(conn)
}
\description{
Validate existing database schema version
}
\keyword{internal}
