% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-clean.R
\name{.derive_emergency}
\alias{.derive_emergency}
\title{Derive emergency certification flag from cert number}
\usage{
.derive_emergency(cert_num)
}
\description{
Derive emergency certification flag from cert number
}
\keyword{internal}
