% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-schema.R
\name{.deserialize_factor_levels}
\alias{.deserialize_factor_levels}
\title{Deserialize factor levels from JSON-like string (without jsonlite)}
\usage{
.deserialize_factor_levels(str)
}
\description{
Deserialize factor levels from JSON-like string (without jsonlite)
}
\keyword{internal}
