% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget-detect.R
\name{.detect_budget_core_columns}
\alias{.detect_budget_core_columns}
\title{Detect core processed classroom-budget identifier columns}
\usage{
.detect_budget_core_columns(col_names)
}
\arguments{
\item{col_names}{Character vector of column names.}
}
\value{
Logical.
}
\description{
Detect core processed classroom-budget identifier columns
}
\keyword{internal}
