% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget-detect.R
\name{.detect_new_markers}
\alias{.detect_new_markers}
\title{Detect new format markers}
\usage{
.detect_new_markers(col_names)
}
\arguments{
\item{col_names}{Character vector of column names.}
}
\value{
Logical.
}
\description{
Detect new format markers
}
\keyword{internal}
