% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage-export.R
\name{.extract_linkage_data}
\alias{.extract_linkage_data}
\title{Extract data frame from linkage objects}
\usage{
.extract_linkage_data(x)
}
\description{
Extract data frame from linkage objects
}
\keyword{internal}
