% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-export.R
\name{.extract_student_data}
\alias{.extract_student_data}
\title{Extract data frame from student clean or panel objects}
\usage{
.extract_student_data(x)
}
\description{
Extract data frame from student clean or panel objects
}
\keyword{internal}
