% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget-read.R
\name{.extract_year_from_bv}
\alias{.extract_year_from_bv}
\title{Extract school year from Budget Version column name}
\usage{
.extract_year_from_bv(bv_col)
}
\arguments{
\item{bv_col}{Character. Budget Version column name.}
}
\value{
Character school year or NA.
}
\description{
Extract school year from Budget Version column name
}
\keyword{internal}
