% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-panel.R
\name{.fill_within_group}
\alias{.fill_within_group}
\title{Fill NA values within a group using non-NA values from the same group}
\usage{
.fill_within_group(x)
}
\description{
Fill NA values within a group using non-NA values from the same group
}
\keyword{internal}
