% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-read.R
\name{.find_classroom_code_col}
\alias{.find_classroom_code_col}
\title{Find the classroom code column in raw data}
\usage{
.find_classroom_code_col(col_names)
}
\arguments{
\item{col_names}{Character vector of column names.}
}
\value{
Column name or NULL.
}
\description{
Find the classroom code column in raw data
}
\keyword{internal}
