% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-read.R
\name{.find_student_id_col}
\alias{.find_student_id_col}
\title{Find the student ID column in raw data}
\usage{
.find_student_id_col(col_names)
}
\arguments{
\item{col_names}{Character vector of column names.}
}
\value{
Column name or NULL.
}
\description{
Find the student ID column in raw data
}
\keyword{internal}
