% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode-detect.R
\name{.geocode_dtype_label}
\alias{.geocode_dtype_label}
\title{Detect R dtype label for a column (character | numeric | logical | other)}
\usage{
.geocode_dtype_label(v)
}
\arguments{
\item{v}{A column vector.}
}
\value{
One of \code{"character"}, \code{"numeric"}, \code{"logical"}, \code{"date"},
\code{"integer"}, \code{"factor"}, \code{"other"}.
}
\description{
Detect R dtype label for a column (character | numeric | logical | other)
}
\keyword{internal}
