% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-clean.R
\name{.get_student_pii_columns}
\alias{.get_student_pii_columns}
\title{Get list of PII columns to remove}
\usage{
.get_student_pii_columns(format)
}
\description{
Get list of PII columns to remove
}
\keyword{internal}
