% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage-coverage.R
\name{.linkage_count_year_rows}
\alias{.linkage_count_year_rows}
\title{Count data rows in selected school years}
\usage{
.linkage_count_year_rows(df, years)
}
\description{
Count data rows in selected school years
}
\keyword{internal}
