% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage-coverage.R
\name{.linkage_geocode_group_pct}
\alias{.linkage_geocode_group_pct}
\title{Build a \code{tibble} of group counts + percentages for a categorical column.}
\usage{
.linkage_geocode_group_pct(x, n_total, col_name)
}
\description{
Build a \code{tibble} of group counts + percentages for a categorical column.
}
\keyword{internal}
