% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage-master.R
\name{.linkage_pct_fmt}
\alias{.linkage_pct_fmt}
\title{Format a percentage that may be NA / NaN for the print method.}
\usage{
.linkage_pct_fmt(x)
}
\description{
Format a percentage that may be NA / NaN for the print method.
}
\keyword{internal}
