% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-clean.R
\name{.order_classroom_columns}
\alias{.order_classroom_columns}
\title{Order columns in logical groups}
\usage{
.order_classroom_columns(df)
}
\description{
Order columns in logical groups
}
\keyword{internal}
