% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-clean.R
\name{.rename_classroom_columns}
\alias{.rename_classroom_columns}
\title{Rename columns using the column map CSV}
\usage{
.rename_classroom_columns(df, column_map)
}
\description{
Rename columns using the column map CSV
}
\keyword{internal}
