% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-clean.R
\name{.rename_student_columns}
\alias{.rename_student_columns}
\title{Rename student columns using CSV column map}
\usage{
.rename_student_columns(df, column_map)
}
\description{
Rename student columns using CSV column map
}
\keyword{internal}
