% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helpers.R
\name{is_yes}
\alias{is_yes}
\title{Evaluate "yes" variants as logical}
\usage{
is_yes(x)
}
\arguments{
\item{x}{Character or logical vector.}
}
\value{
Logical vector.
}
\description{
Evaluate "yes" variants as logical
}
\keyword{internal}
