% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage-export.R
\name{linkage_export_rds}
\alias{linkage_export_rds}
\title{Export Linkage Data to RDS}
\usage{
linkage_export_rds(x, path, compress = TRUE)
}
\arguments{
\item{x}{An \code{alprek_linkage_classroom}, \code{alprek_linkage_student}, or
\code{alprek_linkage_master} object.}

\item{path}{Character. Output file path.}

\item{compress}{Logical. Use compression? Default \code{TRUE}.}
}
\value{
Invisible file path.
}
\description{
Exports linked data to R's native serialized format.
}
