% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage-master.R
\name{linkage_summary_stats}
\alias{linkage_summary_stats}
\title{Summary Statistics for Linkage Data}
\usage{
linkage_summary_stats(x, by = "school_year")
}
\arguments{
\item{x}{An \code{alprek_linkage_master}, \code{alprek_linkage_classroom}, or
\code{alprek_linkage_student} object.}

\item{by}{Character. Grouping variable. Default \code{"school_year"}.}
}
\value{
A tibble of summary statistics.
}
\description{
Computes summary statistics for linked data by school year.
}
