% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications-clean.R
\name{print.alprek_applications_clean}
\alias{print.alprek_applications_clean}
\title{Print method for alprek_applications_clean}
\usage{
\method{print}{alprek_applications_clean}(x, ...)
}
\arguments{
\item{x}{An alprek_applications_clean object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_applications_clean
}
