% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications-config.R
\name{print.alprek_applications_config}
\alias{print.alprek_applications_config}
\title{Print method for alprek_applications_config}
\usage{
\method{print}{alprek_applications_config}(x, ...)
}
\arguments{
\item{x}{An \code{alprek_applications_config} object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_applications_config
}
