% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage-applications.R
\name{print.alprek_applications_linkage}
\alias{print.alprek_applications_linkage}
\title{Print method for \code{alprek_applications_linkage}}
\usage{
\method{print}{alprek_applications_linkage}(x, ...)
}
\arguments{
\item{x}{An \code{alprek_applications_linkage} object.}

\item{...}{Ignored.}
}
\description{
Print method for \code{alprek_applications_linkage}
}
