% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications-transform.R
\name{print.alprek_applications_master}
\alias{print.alprek_applications_master}
\title{Print method for \code{alprek_applications_master}}
\usage{
\method{print}{alprek_applications_master}(x, ...)
}
\arguments{
\item{x}{An \code{alprek_applications_master} object.}

\item{...}{Ignored.}
}
\description{
Print method for \code{alprek_applications_master}
}
