% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications-panel.R
\name{print.alprek_applications_panel}
\alias{print.alprek_applications_panel}
\title{Print method for \code{alprek_applications_panel}}
\usage{
\method{print}{alprek_applications_panel}(x, ...)
}
\arguments{
\item{x}{An \code{alprek_applications_panel} object.}

\item{...}{Ignored.}
}
\description{
Print method for \code{alprek_applications_panel}
}
