% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications-read.R
\name{print.alprek_applications_raw}
\alias{print.alprek_applications_raw}
\title{Print method for alprek_applications_raw}
\usage{
\method{print}{alprek_applications_raw}(x, ...)
}
\arguments{
\item{x}{An alprek_applications_raw object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_applications_raw
}
