% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications-reconcile.R
\name{print.alprek_applications_reconciled}
\alias{print.alprek_applications_reconciled}
\title{Print method for \code{alprek_applications_reconciled}}
\usage{
\method{print}{alprek_applications_reconciled}(x, ...)
}
\arguments{
\item{x}{An \code{alprek_applications_reconciled} object.}

\item{...}{Ignored.}
}
\description{
Print method for \code{alprek_applications_reconciled}
}
