% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications-validate.R
\name{print.alprek_applications_validation}
\alias{print.alprek_applications_validation}
\title{Print method for \code{alprek_applications_validation}}
\usage{
\method{print}{alprek_applications_validation}(x, ...)
}
\arguments{
\item{x}{An \code{alprek_applications_validation} object.}

\item{...}{Ignored.}
}
\description{
Print method for \code{alprek_applications_validation}
}
