% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget-config.R
\name{print.alprek_budget_config}
\alias{print.alprek_budget_config}
\title{Print method for alprek_budget_config}
\usage{
\method{print}{alprek_budget_config}(x, ...)
}
\arguments{
\item{x}{An alprek_budget_config object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_budget_config
}
