% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget-clean.R
\name{print.alprek_budget_long}
\alias{print.alprek_budget_long}
\title{Print method for alprek_budget_long}
\usage{
\method{print}{alprek_budget_long}(x, ...)
}
\arguments{
\item{x}{An alprek_budget_long object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_budget_long
}
