% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget-transform.R
\name{print.alprek_budget_master}
\alias{print.alprek_budget_master}
\title{Print method for alprek_budget_master}
\usage{
\method{print}{alprek_budget_master}(x, ...)
}
\arguments{
\item{x}{An alprek_budget_master object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_budget_master
}
