% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget-master.R
\name{print.alprek_budget_panel}
\alias{print.alprek_budget_panel}
\title{Print method for alprek_budget_panel}
\usage{
\method{print}{alprek_budget_panel}(x, ...)
}
\arguments{
\item{x}{An alprek_budget_panel object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_budget_panel
}
