% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget-read.R
\name{print.alprek_budget_raw}
\alias{print.alprek_budget_raw}
\title{Print method for alprek_budget_raw}
\usage{
\method{print}{alprek_budget_raw}(x, ...)
}
\arguments{
\item{x}{An alprek_budget_raw object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_budget_raw
}
