% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget-validate.R
\name{print.alprek_budget_validation}
\alias{print.alprek_budget_validation}
\title{Print method for alprek_budget_validation}
\usage{
\method{print}{alprek_budget_validation}(x, ...)
}
\arguments{
\item{x}{An alprek_budget_validation object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_budget_validation
}
