% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-clean.R
\name{print.alprek_classroom_clean}
\alias{print.alprek_classroom_clean}
\title{Print method for alprek_classroom_clean}
\usage{
\method{print}{alprek_classroom_clean}(x, ...)
}
\arguments{
\item{x}{An alprek_classroom_clean object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_classroom_clean
}
