% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-config.R
\name{print.alprek_classroom_config}
\alias{print.alprek_classroom_config}
\title{Print method for alprek_classroom_config}
\usage{
\method{print}{alprek_classroom_config}(x, ...)
}
\arguments{
\item{x}{An alprek_classroom_config object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_classroom_config
}
