% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-panel.R
\name{print.alprek_classroom_panel}
\alias{print.alprek_classroom_panel}
\title{Print method for alprek_classroom_panel}
\usage{
\method{print}{alprek_classroom_panel}(x, ...)
}
\arguments{
\item{x}{An alprek_classroom_panel object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_classroom_panel
}
