% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-read.R
\name{print.alprek_classroom_raw}
\alias{print.alprek_classroom_raw}
\title{Print method for alprek_classroom_raw}
\usage{
\method{print}{alprek_classroom_raw}(x, ...)
}
\arguments{
\item{x}{An alprek_classroom_raw object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_classroom_raw
}
