% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-validate.R
\name{print.alprek_classroom_validation}
\alias{print.alprek_classroom_validation}
\title{Print method for alprek_classroom_validation}
\usage{
\method{print}{alprek_classroom_validation}(x, ...)
}
\arguments{
\item{x}{An alprek_classroom_validation object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_classroom_validation
}
