% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode-detect.R
\name{print.alprek_geocode_delivery_diff}
\alias{print.alprek_geocode_delivery_diff}
\title{Print method for alprek_geocode_delivery_diff}
\usage{
\method{print}{alprek_geocode_delivery_diff}(x, ...)
}
\arguments{
\item{x}{An \code{alprek_geocode_delivery_diff} object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_geocode_delivery_diff
}
