% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage-geocode.R
\name{print.alprek_geocode_linkage_applications}
\alias{print.alprek_geocode_linkage_applications}
\title{Print method for \code{alprek_geocode_linkage_applications}}
\usage{
\method{print}{alprek_geocode_linkage_applications}(x, ...)
}
\arguments{
\item{x}{An \code{alprek_geocode_linkage_applications} object.}

\item{...}{Ignored.}
}
\description{
Print method for \code{alprek_geocode_linkage_applications}
}
