% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage-geocode.R
\name{print.alprek_geocode_linkage_classroom}
\alias{print.alprek_geocode_linkage_classroom}
\title{Print method for \code{alprek_geocode_linkage_classroom}}
\usage{
\method{print}{alprek_geocode_linkage_classroom}(x, ...)
}
\arguments{
\item{x}{An \code{alprek_geocode_linkage_classroom} object.}

\item{...}{Ignored.}
}
\description{
Print method for \code{alprek_geocode_linkage_classroom}
}
