% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode-transform.R
\name{print.alprek_geocode_master}
\alias{print.alprek_geocode_master}
\title{Print method for \code{alprek_geocode_master}}
\usage{
\method{print}{alprek_geocode_master}(x, ...)
}
\arguments{
\item{x}{An \code{alprek_geocode_master} object.}

\item{...}{Ignored.}
}
\description{
Print method for \code{alprek_geocode_master}
}
