% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode-panel.R
\name{print.alprek_geocode_panel}
\alias{print.alprek_geocode_panel}
\title{Print method for \code{alprek_geocode_panel}}
\usage{
\method{print}{alprek_geocode_panel}(x, ...)
}
\arguments{
\item{x}{An \code{alprek_geocode_panel} object.}

\item{...}{Ignored.}
}
\description{
Print method for \code{alprek_geocode_panel}
}
