% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode-read.R
\name{print.alprek_geocode_raw}
\alias{print.alprek_geocode_raw}
\title{Print method for alprek_geocode_raw}
\usage{
\method{print}{alprek_geocode_raw}(x, ...)
}
\arguments{
\item{x}{An \code{alprek_geocode_raw} object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_geocode_raw
}
