% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode-reconcile.R
\name{print.alprek_geocode_reconciled}
\alias{print.alprek_geocode_reconciled}
\title{Print method for \code{alprek_geocode_reconciled}}
\usage{
\method{print}{alprek_geocode_reconciled}(x, ...)
}
\arguments{
\item{x}{An \code{alprek_geocode_reconciled} object.}

\item{...}{Ignored.}
}
\description{
Print method for \code{alprek_geocode_reconciled}
}
