% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode-validate.R
\name{print.alprek_geocode_validation}
\alias{print.alprek_geocode_validation}
\title{Print method for \code{alprek_geocode_validation}}
\usage{
\method{print}{alprek_geocode_validation}(x, ...)
}
\arguments{
\item{x}{An \code{alprek_geocode_validation} object.}

\item{...}{Ignored.}
}
\description{
Print method for \code{alprek_geocode_validation}
}
