% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage-join.R
\name{print.alprek_linkage_classroom}
\alias{print.alprek_linkage_classroom}
\title{Print method for alprek_linkage_classroom}
\usage{
\method{print}{alprek_linkage_classroom}(x, ...)
}
\arguments{
\item{x}{An alprek_linkage_classroom object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_linkage_classroom
}
