% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage-master.R
\name{print.alprek_linkage_master}
\alias{print.alprek_linkage_master}
\title{Print method for alprek_linkage_master}
\usage{
\method{print}{alprek_linkage_master}(x, ...)
}
\arguments{
\item{x}{An alprek_linkage_master object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_linkage_master
}
