% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage-join.R
\name{print.alprek_linkage_student}
\alias{print.alprek_linkage_student}
\title{Print method for alprek_linkage_student}
\usage{
\method{print}{alprek_linkage_student}(x, ...)
}
\arguments{
\item{x}{An alprek_linkage_student object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_linkage_student
}
