% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage-diagnostics.R
\name{print.alprek_linkage_validation}
\alias{print.alprek_linkage_validation}
\title{Print method for alprek_linkage_validation}
\usage{
\method{print}{alprek_linkage_validation}(x, ...)
}
\arguments{
\item{x}{An alprek_linkage_validation object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_linkage_validation
}
