% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-clean.R
\name{print.alprek_student_clean}
\alias{print.alprek_student_clean}
\title{Print method for alprek_student_clean}
\usage{
\method{print}{alprek_student_clean}(x, ...)
}
\arguments{
\item{x}{An alprek_student_clean object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_student_clean
}
