% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-config.R
\name{print.alprek_student_config}
\alias{print.alprek_student_config}
\title{Print method for alprek_student_config}
\usage{
\method{print}{alprek_student_config}(x, ...)
}
\arguments{
\item{x}{An alprek_student_config object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_student_config
}
