% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-panel.R
\name{print.alprek_student_panel}
\alias{print.alprek_student_panel}
\title{Print method for alprek_student_panel}
\usage{
\method{print}{alprek_student_panel}(x, ...)
}
\arguments{
\item{x}{An alprek_student_panel object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_student_panel
}
