% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-read.R
\name{print.alprek_student_raw}
\alias{print.alprek_student_raw}
\title{Print method for alprek_student_raw}
\usage{
\method{print}{alprek_student_raw}(x, ...)
}
\arguments{
\item{x}{An alprek_student_raw object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_student_raw
}
