% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-validate.R
\name{print.alprek_student_validation}
\alias{print.alprek_student_validation}
\title{Print method for alprek_student_validation}
\usage{
\method{print}{alprek_student_validation}(x, ...)
}
\arguments{
\item{x}{An alprek_student_validation object.}

\item{...}{Ignored.}
}
\description{
Print method for alprek_student_validation
}
