% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-export.R
\name{student_export_csv}
\alias{student_export_csv}
\title{Export Student Data to CSV}
\usage{
student_export_csv(x, path)
}
\arguments{
\item{x}{An \code{alprek_student_clean} or \code{alprek_student_panel} object.}

\item{path}{Character. Output file path.}
}
\value{
Invisible file path of the written file.
}
\description{
Exports processed student data to CSV format.
}
