% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-export.R
\name{student_export_rds}
\alias{student_export_rds}
\title{Export Student Data to RDS}
\usage{
student_export_rds(x, path, compress = TRUE)
}
\arguments{
\item{x}{An \code{alprek_student_clean} or \code{alprek_student_panel} object.}

\item{path}{Character. Output file path.}

\item{compress}{Logical. Use compression? Default \code{TRUE}.}
}
\value{
Invisible file path.
}
\description{
Exports processed student data to R's native serialized format.
}
